refactor(input.tsx): add faker and fix input test
