+++ src/components/organisms/sections/counting-section.tsx
@@ -1,7 +1,7 @@
 import { type FC, useState } from 'react'
 
 import Datepicker, { type DateValueType } from '@atoms/datepicker/datepicker'
-import Input from '@atoms/input'
+import Input from '@atoms/input/input'
 
 import useAuthStore from '@store/use-auth-store'
